Model climate response at individual sites
Reviewer 3 insists that regional analyses are not valid and we need to show individual site responses.  But there's not enough power at individual sites and it ends up focusing on anaomalies rather than the role of absolute temperature or precipitation values.  We believe reviewer 3 is being short-sited both in spatial and temporal scope.