Ecology R1 revisions -- step 1:
Paper is apparenlty not going to fly if we don't have random effects and is "too complicated."   To satisfy reviewers, we will no longer borrow strenght across species to capture site dynaics and will just do separate models for each species.  This allows the random effects (random intercepts) to work.  (Not my preferrred route, but whatever.)  In doing this, most of the variance decomposition framework that was a relic of Ecology Letters is no longer relevant, which should allow us to greatly reduce that part of the paper and the scripts.  We'll see how it goes.

Lots more commites to come, but this is the first step